src/library/requirements.txt
# preprocessor
pillow==11.3.0
numpy==2.2.6
# json_to_table
openpyxl==3.1.5
# craft
torch==2.8.0
torchvision==0.23.0
opencv-python-headless==4.12.0.88
# paddle
paddlepaddle==3.2.0
paddleocr==3.2.0